+++ src/main/java/racecar/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package racecar;
 
 import java.util.ArrayList;
 import java.util.List;
