+++ src/test/java/racecar/CarTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racecar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
